pyproject.toml
[project]
name = "GraphQLer"
version = "2.3.8"
description = "A cutting edge context aware GraphQL API fuzzing tool!"
authors = [{ name = "Omar2535", email = "omar@9tsai.ca" }]
requires-python = ">=3.12,<3.13"
readme = "README.md"
license = "MIT"
dependencies = [
    "requests>=2.32.5",
    "urllib3>=2.6.3",
    "graphql-core>=3.2.8",
    "pyyaml>=6.0.3",
    "levenshtein>=0.27.3",
    "clairvoyance>=2.5.5",
    "cloudpickle>=3.1.2,<4",
    "matplotlib>=3.8.0",
    "networkx>=3.6.1",
    "litellm>=1.67.0",
    "websockets>=14.0",
    "textual>=6.0.0,<8.0.0",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0.0", "fastmcp>=3.1.1"]

[project.scripts]
graphqler-mcp = "graphqler.utils.mcp_utils.server:serve"

[tool.ruff]
line-length = 180

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock",
    "pytest-xdist>=3.0.0",
    "flake8>=7.3.0",
    "pre-commit>=4.5.1",
    "ruff>=0.15.7",
    "mkdocs>=1.6.1,<2",
    "lark>=1.3.1",
    "ty>=0.0.25",
]

[tool.ty.src]
exclude = ["sample-graphql-apis/"]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
